+++ src/main/java/org/usfirst/frc4048/common/autochooser/FieldLocation.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc4048.common.auto;
+package org.usfirst.frc4048.common.autochooser;
 public enum FieldLocation {
      Right,
      Middle,
