+++ src/main/java/org/usfirst/frc4048/common/autochooser/PlaceHolderCommand.java
@@ -1,10 +1,7 @@
-package org.usfirst.frc4048.common.auto;
+package org.usfirst.frc4048.common.autochooser;
 
-import edu.wpi.first.util.function.BooleanConsumer;
 import edu.wpi.first.wpilibj2.command.*;
 
-import java.util.Set;
-import java.util.function.BooleanSupplier;
 //Why is CommandBase abstract if it has not abstract methods????
 public class PlaceHolderCommand extends CommandBase{
