+++ src/main/java/org/usfirst/frc4048/common/autochooser/chooser/AutoChooser.java
@@ -1,0 +1,10 @@
+package org.usfirst.frc4048.common.autochooser.chooser;
+
+import edu.wpi.first.wpilibj2.command.Command;
+import frc.robot.auto.event.AutoEventProvider;
+
+public interface AutoChooser {
+     Command getAutoCommand();
+     AutoEventProvider getAutoEventProvider();
+
+}
